seperation objects et rectification imports
