+++ Packing1D/__init__.py
@@ -1,5 +1,5 @@
-from bacs import Bac1D
-from objects import PackingObject1D
+from bacs.Bac1D import Bac1D
+from objects.PackingObject1D import PackingObject1D
 from packingException import IncompatibleBacException
 def worstFit1D( B:int , objects:list):
     '''
@@ -16,6 +16,7 @@ def worstFit1D( B:int , objects:list):
     num_bac = 1
     bacs.append(Bac1D(B,num_bac))
     for packing_object in objects:
+        bacs = sorted (bacs,key=lambda bac : bac.get_free_space(),reverse=True)
         added = False
         for bac in bacs:
             try:
